requirements.txt
torch>=2.0.0
transformers>=4.40.0
vllm>=0.4.0
accelerate>=0.30.0
datasets>=2.16.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0
sentencepiece>=0.1.99
bitsandbytes>=0.41.0
autoawq>=0.2.0
rank-bm25>=0.2.2
einops>=0.7.0


